+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/action/swf1/ActionPlay.java
@@ -14,13 +14,16 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library.
  */
-package com.jpexs.decompiler.flash.action.swf3;
+package com.jpexs.decompiler.flash.action.swf1;
 
 import com.jpexs.decompiler.flash.action.Action;
 import com.jpexs.decompiler.flash.action.DisplayObject;
 import com.jpexs.decompiler.flash.action.LocalDataArea;
 import com.jpexs.decompiler.flash.action.as2.Trait;
-import com.jpexs.decompiler.flash.action.model.NextFrameActionItem;
+import com.jpexs.decompiler.flash.action.model.DirectValueActionItem;
+import com.jpexs.decompiler.flash.action.model.GotoFrame2ActionItem;
+import com.jpexs.decompiler.flash.action.model.GotoFrameActionItem;
+import com.jpexs.decompiler.flash.action.model.PlayActionItem;
 import com.jpexs.decompiler.flash.types.annotations.SWFVersion;
 import com.jpexs.decompiler.graph.GraphSourceItem;
 import com.jpexs.decompiler.graph.GraphTargetItem;
@@ -33,37 +36,38 @@ import java.util.Map;
 import java.util.Set;
 
 /**
- * NextFrame action - Jumps to the next frame in the current timeline.
+ * Play action - Plays the current timeline.
  *
  * @author JPEXS
  */
-@SWFVersion(from = 3)
-public class ActionNextFrame extends Action {
+@SWFVersion(from = 1)
+public class ActionPlay extends Action {
 
     /**
      * Constructor
      */
-    public ActionNextFrame() {
-        super(0x04, 0, Utf8Helper.charsetName);
+    public ActionPlay() {
+        super(0x06, 0, Utf8Helper.charsetName);
     }
 
     @Override
     public String toString() {
-        return "NextFrame";
+        return "Play";
     }
 
     @Override
     public boolean execute(LocalDataArea lda) {
-        int f = ((DisplayObject) lda.target).getCurrentFrame();
-        if (f < ((DisplayObject) lda.target).getTotalFrames()) {
-            ((DisplayObject) lda.target).gotoFrame(f + 1);
-        }
-
+        ((DisplayObject) lda.target).play();
         return true;
     }
 
     @Override
     public void translate(Set<String> usedDeobfuscations, Map<String, Map<String, Trait>> uninitializedClassTraits, SecondPassData secondPassData, boolean insideDoInitAction, GraphSourceItem lineStartAction, TranslateStack stack, List<GraphTargetItem> output, HashMap<Integer, String> regNames, HashMap<String, GraphTargetItem> variables, HashMap<String, GraphTargetItem> functions, int staticOperation, String path) {
-        output.add(new NextFrameActionItem(this, lineStartAction));
+        if (!output.isEmpty() && (output.get(output.size() - 1) instanceof GotoFrameActionItem)) {
+            GotoFrameActionItem gta = (GotoFrameActionItem) output.remove(output.size() - 1);
+            output.add(new GotoFrame2ActionItem(this, lineStartAction, new DirectValueActionItem(gta.frame + 1), false, true, 0));
+        } else {
+            output.add(new PlayActionItem(this, lineStartAction));
+        }
     }
 }