+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/action/swf1/ActionPrevFrame.java
@@ -14,12 +14,13 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library.
  */
-package com.jpexs.decompiler.flash.action.swf3;
+package com.jpexs.decompiler.flash.action.swf1;
 
 import com.jpexs.decompiler.flash.action.Action;
+import com.jpexs.decompiler.flash.action.DisplayObject;
 import com.jpexs.decompiler.flash.action.LocalDataArea;
 import com.jpexs.decompiler.flash.action.as2.Trait;
-import com.jpexs.decompiler.flash.action.model.StopAllSoundsActionItem;
+import com.jpexs.decompiler.flash.action.model.PrevFrameActionItem;
 import com.jpexs.decompiler.flash.types.annotations.SWFVersion;
 import com.jpexs.decompiler.graph.GraphSourceItem;
 import com.jpexs.decompiler.graph.GraphTargetItem;
@@ -32,33 +33,37 @@ import java.util.Map;
 import java.util.Set;
 
 /**
- * StopSounds action - Stops all sounds.
+ * PrevFrame action - Jumps to the previous frame in the current timeline.
  *
  * @author JPEXS
  */
-@SWFVersion(from = 3)
-public class ActionStopSounds extends Action {
+@SWFVersion(from = 1)
+public class ActionPrevFrame extends Action {
 
     /**
      * Constructor
      */
-    public ActionStopSounds() {
-        super(0x09, 0, Utf8Helper.charsetName);
+    public ActionPrevFrame() {
+        super(0x05, 0, Utf8Helper.charsetName);
     }
 
     @Override
     public String toString() {
-        return "StopSounds";
+        return "PrevFrame";
     }
 
     @Override
     public boolean execute(LocalDataArea lda) {
-        lda.stage.stopSounds();
+        int f = ((DisplayObject) lda.target).getCurrentFrame();
+        if (f > 1) {
+            ((DisplayObject) lda.target).gotoFrame(f - 1);
+        }
+
         return true;
     }
 
     @Override
     public void translate(Set<String> usedDeobfuscations, Map<String, Map<String, Trait>> uninitializedClassTraits, SecondPassData secondPassData, boolean insideDoInitAction, GraphSourceItem lineStartAction, TranslateStack stack, List<GraphTargetItem> output, HashMap<Integer, String> regNames, HashMap<String, GraphTargetItem> variables, HashMap<String, GraphTargetItem> functions, int staticOperation, String path) {
-        output.add(new StopAllSoundsActionItem(this, lineStartAction));
+        output.add(new PrevFrameActionItem(this, lineStartAction));
     }
 }