+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/action/swf1/ActionToggleQuality.java
@@ -14,13 +14,12 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library.
  */
-package com.jpexs.decompiler.flash.action.swf3;
+package com.jpexs.decompiler.flash.action.swf1;
 
 import com.jpexs.decompiler.flash.action.Action;
-import com.jpexs.decompiler.flash.action.DisplayObject;
 import com.jpexs.decompiler.flash.action.LocalDataArea;
 import com.jpexs.decompiler.flash.action.as2.Trait;
-import com.jpexs.decompiler.flash.action.model.PrevFrameActionItem;
+import com.jpexs.decompiler.flash.action.model.ToggleHighQualityActionItem;
 import com.jpexs.decompiler.flash.types.annotations.SWFVersion;
 import com.jpexs.decompiler.graph.GraphSourceItem;
 import com.jpexs.decompiler.graph.GraphTargetItem;
@@ -33,37 +32,33 @@ import java.util.Map;
 import java.util.Set;
 
 /**
- * PrevFrame action - Jumps to the previous frame in the current timeline.
+ * ToggleQuality action - Toggles the quality of the Flash Player.
  *
  * @author JPEXS
  */
-@SWFVersion(from = 3)
-public class ActionPrevFrame extends Action {
+@SWFVersion(from = 1)
+public class ActionToggleQuality extends Action {
 
     /**
      * Constructor
      */
-    public ActionPrevFrame() {
-        super(0x05, 0, Utf8Helper.charsetName);
+    public ActionToggleQuality() {
+        super(0x08, 0, Utf8Helper.charsetName);
     }
 
     @Override
     public String toString() {
-        return "PrevFrame";
+        return "ToggleQuality";
     }
 
     @Override
     public boolean execute(LocalDataArea lda) {
-        int f = ((DisplayObject) lda.target).getCurrentFrame();
-        if (f > 1) {
-            ((DisplayObject) lda.target).gotoFrame(f - 1);
-        }
-
+        lda.stage.toggleQuality();
         return true;
     }
 
     @Override
     public void translate(Set<String> usedDeobfuscations, Map<String, Map<String, Trait>> uninitializedClassTraits, SecondPassData secondPassData, boolean insideDoInitAction, GraphSourceItem lineStartAction, TranslateStack stack, List<GraphTargetItem> output, HashMap<Integer, String> regNames, HashMap<String, GraphTargetItem> variables, HashMap<String, GraphTargetItem> functions, int staticOperation, String path) {
-        output.add(new PrevFrameActionItem(this, lineStartAction));
+        output.add(new ToggleHighQualityActionItem(this, lineStartAction));
     }
 }