+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/action/swf1/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * SWF 1 actions.
+ */
+package com.jpexs.decompiler.flash.action.swf1;